Add a .clang-tidy configuration file
Clang-tidy is a tool to provide additional static analysis warningts. It makes sense to keep as many warnings as possible active. Warning that are not usefull in scope of this project are disabled with a rationale.
The latest version of Visual Studio 2019 can also use clang-tidy to perform background checking.

+++ src/interface.cpp
@@ -116,7 +116,7 @@ jpegls_errc JpegLsEncodeStream(const ByteStreamInfo destination, size_t& bytesWr
         }
         else if (info.bitsPerSample > 12)
         {
-            const auto default_preset_coding_parameters{compute_default((1 << info.bitsPerSample) - 1, info.allowedLossyError)};
+            const auto default_preset_coding_parameters{compute_default((1U << static_cast<uint32_t>(info.bitsPerSample)) - 1, info.allowedLossyError)};
             writer.WriteJpegLSPresetParametersSegment(default_preset_coding_parameters);
         }
